refactor WalletConnectCosignerUtils
